Update test cases for netty
